:dancer: Organise tests files more closely to app files
