+++ test/test-lib/resources/syndication_responses/paged_pharmacies_postcode_search.xml
