+++ test/test-lib/resources/syndication_responses/single_pharmacy_overview.xml
